refactor: 배경색 없애기 및 box 커스텀 컴포넌트사용, Divider 사용
