+++ src/components/modal/priceAlertModal/index.tsx
