+++ src/components/priceAlertSetting/styles.tsx
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import { Box } from '@mui/material';
+import Box from '../box/styles';
 
 export const Container = styled(Box)`
   display: flex;
