+++ src/pages/search/initial/index.tsx
@@ -3,6 +3,8 @@ import { useNavigate } from 'react-router-dom';
 import PopularSearch from '../../../components/searches/popularSearches';
 import RecentSearches from '../../../components/searches/recentSearches';
 import RecommendSearch from '../../../components/searches/recommendSearches';
+import Divider from '@mui/material/Divider';
+import Box from '../../../components/box/styles';
 
 type SearchInitialPageProps = {
   searchValue: string;
@@ -25,12 +27,30 @@ const SearchInitialPage: React.FC<SearchInitialPageProps> = ({
   return (
     <>
       {searchValue ? (
-        <RecommendSearch onSearch={handleSearch} />
+        <Box>
+          <RecommendSearch onSearch={handleSearch} />
+          <Divider
+            sx={{
+              my: 2,
+              borderColor: 'rgba(0, 0, 0, 0.1)', // 색상 연하게 설정
+              width: '80%', // 길이 80%로 설정
+              mx: 'auto' // 가운데 정렬
+            }}
+          />
+        </Box>
       ) : (
-        <>
+        <Box>
           <RecentSearches />
+          <Divider
+            sx={{
+              my: 2,
+              borderColor: 'rgba(0, 0, 0, 0.1)', // 색상 연하게 설정
+              width: '95%', // 길이 80%로 설정
+              mx: 'auto' // 가운데 정렬
+            }}
+          />
           <PopularSearch />
-        </>
+        </Box>
       )}
     </>
   );
